HARMONY-1426
Array of monitors created by function GetOwnedMonitorsInfo can be freed with Deallocate JVMTI function. This array is 
taken from Thread Manager by calling jthread_get_owned_monitors. But Thread Manager uses APR pools for allocating 
memory for this array. So when JVMTI agent tries to use Deallocate on the array pointer the program usually crashes 
since the pointer was not allocated by JVMTI Allocate.

Inside of JVMTI implementation Allocate simply calls malloc. So to solve this prooblem it is enough to change Thread 
Manager allocator to malloc. I've done this fix in the same way as other functions in this file seem to be fixed.

Ugly tabs formatting is not fixed.
 



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@446791 13f79535-47bb-0310-9956-ffa450edef68