HARMONY-1541
The current implementation of JVMTI breakpoints invokes agent from inside of exception (on Windows) or signal (on 
Linux) handler. This usually works but may cause problems if for example agent decides to stop the thread with 
breakpoint for some long time. On Windows exception handler cannot be invoked by another thread until other exception 
handler is running. This may cause dead locks with any exception happening in other threads, e.g. C++ exceptions on 
windows are done through the same mechanism. On Linux calling arbitrary functions from inside of signal handler may 
not work as well.

The solution for this problem is if JVMTI sees that happened breakpoint is actually set by JVMTI, for processing it 
it should exit the handler. In order to start breakpoint processing the handler exit IP is changed from the IP 
pointing to the interrupted location to the JVMTI processing handler. Then exception/signal handler exits and makes 
the kernel happy while execution is transferred to JVMTI breakpoints handler. This also is done while holding 
breakpoints list lock so that no intermittent modification, like removing the breakpoin which caused exception is 
possible.

Ubuntu 6 - smoke, c-unit, kernel



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@449228 13f79535-47bb-0310-9956-ffa450edef68