HARMONY-1545
I've reworked the implementation of JVMTI breakpoints and single step. The code which sends breakpoint and single 
step events currently is ugly and uses unobvious logic. In the new implementation the infrastructure is much more 
clean and clear :)

Main class is VMBreakPoints. It is the container for all breakpoint instrumentation and handler for all breakpoint 
events. A VMBreakInterface is a class for breakpoints references container. It can be used for any purpose, be it 
SingleStep or user breakpoints container. In VMBreakInterface breakpoints instrumentations are referenced by 
VMBreakPointRef.

VMBreakPoints has handlers for native (JIT mode) and interpreter mode breakpoints. When such handler is called, the 
processing goes through all registered VMBreakInterface instances according to their priorities (currently two 
priorities are supported - SingleStep is first according to JVMTI spec about colocated events) and calls 
VMBreakInterface handling callback which can send an event of type which this interface supports. 


Ubuntu 6 - smoke, c-unit



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@449232 13f79535-47bb-0310-9956-ffa450edef68