HARMONY-1548
I like the new infrastructure, contained in Harmony-1545, however I found
several bugs in the implementation.

- Improved insterting breakpoint interface.
- Improved trace for inserting breakpoints.
- Fixed declaration of VMBreakPoints class.
- Improved interface for breakpoints

I also added a lot of useful tracing to debug this code.

This patch also contains several fixes in parts of code unrelated to the new
infrastructure but necessary for correct SingleStep and breakpoints
functionality.

- Added check of existence of pending breakpoints after lock.
- Fixed setting single step breakpoint for Linux
- Fixed erroneous breakpoint events reporting 

Ubuntu 6 - smoke, c-unit and kernel



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@449237 13f79535-47bb-0310-9956-ffa450edef68