HARMONY-1613
In interpreter mode function GetStackTrace() misses frame with main() function.
The problem is in skip_old_frames() function in interpreter. It reports wrong additional missing frame.

Wasn't able to test this specific fix, but it didn't break anything.  Can we have a test to show it?

Ubuntu 6, smoke, c-unit and kernel (as they are)



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@452152 13f79535-47bb-0310-9956-ffa450edef68