HARMONY-1598
FramePop event is not reported in JIT mode.

The main problem is in jvmtiNotifyFramePop() function. It always reports JVMTI_ERROR_OPAQUE_FRAME error in JIT mode. 
Because of this "Step Over" doesn't work in Eclipse debug.

Besides that prediction for the next bytecode for instruction return worked incorrectly. IP was set not on the 
instruction after invoke, but on the following instruction.

Ubuntu 6, smoke, c-unit and kernel (whateverworked)

We need some JVMTI tests :)



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@452454 13f79535-47bb-0310-9956-ffa450edef68