HARMONY-1651
Mostly - rejected the catch(NPE) and did something to be explicit
about what we are doing when there is no specVersion.

All three tests pass.  Note, I had to also add a manifest to the
kernel.jar for this to work - not sure how this worked without it.

All three tests pass, and nothing else seems to have broken.

Ubuntu 6, smoke, c-unit, ~kernel



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@453042 13f79535-47bb-0310-9956-ffa450edef68