HARMONY-1571
This is one line compilation fix for a warning of "creating pointer from an integer of different size" for x86_64 
platforms. The JVMTI code is still meant to work only on ia32 but for x86_64 it should compile, so this patch is a 
workaround to allow compilation.

Ubuntu 6 - 32-bit - please test _64

smoke, c-unit, ~kernel



git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@453428 13f79535-47bb-0310-9956-ffa450edef68