HARMONY-1561
Interpreter throws IllegalAccessError on putfield opcode for all methods except <init>.
It should not throw the error if the <init> method for this object is in the calling stack.
Reproducible on activemq-4.0.2.

Had to -R putfield.patch before applying H-1561.diff

TestPutFinal.j passes , as do smoke, c-unit, kernel on Ubuntu 6




git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/drlvm/trunk@463912 13f79535-47bb-0310-9956-ffa450edef68