Adding formal, platform specific exclusion lists for the kernel tests (hey, it's a start).
As part of this, I "reach over" and use the classlib properties.xml
so that the defined platform, arch and vm terms are the same
as used in classlib.  These are different than the set used 
in DRLVM, but we have to converge somehow.

The excludes initially seeded are the ones that were buried in the 
kernel.test.xml file, which are now not used



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@477741 13f79535-47bb-0310-9956-ffa450edef68