Applied HARMONY-2721 [drlvm][test] Add test in HARMONY-2083 to regression test stuite
It appears that there is a race condition in using class_report_failure because
it is called after Class is unlocked. So it is possible that 2 threads come
with different errors to calling cp.resolve_as_error. To avoid errors
duplication since it is necessary to use only the first error it is better
to check for existing error before calling cp.resolve_as_error.



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@499839 13f79535-47bb-0310-9956-ffa450edef68