Fixed bug HARMONY-3758 [drlvm][classloader] assert fires in String_Pool::lookup()
For some reason in unaligned case when string is read byte by byte the signed char bytes
were converted to POINTER_SIZE_INT with sign extend. When string contains unicode
characters ( >= 128) this leads to incorrect results because upper bytes of the
byte are filled with 0xff bytes. The fix is to consider bytes as unsigned before
extending them to POINTER_SIZE_INT.



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@538982 13f79535-47bb-0310-9956-ffa450edef68