Committed fix for HARMONY-4533
With lazy resolution it is not necessary that an invokevirtual or invokeinterface
bytecodes are always complied to call the method. When the method is not yet resolved
it is possible that the call points to method's resolution stub. In this case
single stepping should begin when the method is ready which happens automatically
in the current design of single step. So nothing should be done except that the
assertion is no longer true.



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@569108 13f79535-47bb-0310-9956-ffa450edef68